typeck/expr.rs: move check_expr_struct here.
